Resolve merge conflicts with 553144 changes from sprint4.5 & applied reviewer suggestions on PR107
