+++ src/EPR.RegulatorService.Facade.API/appsettings.json
@@ -127,8 +127,7 @@
       "RegistrationFeeByRegistrationMaterialId": "api/v{0}/registrationMaterials/{1}/paymentFees",
       "MarkAsDulyMadeByRegistrationMaterialId": "api/v{0}/registrationMaterials/{1}/markAsDulyMade",
       "RegistrationAccreditationReference": "api/v{0}/registrationMaterials/{1}/RegistrationAccreditationReference",
-      "RegistrationByIdWithAccreditations": "api/v{0}/registrations/{1}/accreditations"
-      "RegistrationAccreditationReference": "api/v{0}/registrationMaterials/{1}/RegistrationAccreditationReference",
+      "RegistrationByIdWithAccreditations": "api/v{0}/registrations/{1}/accreditations",      
       "AccreditationFeeByRegistrationMaterialId": "api/v{0}/accreditationMaterials/{1}/paymentFees",
       "AccreditationMarkAsDulyMadeByRegistrationMaterialId": "api/v{0}/accreditationMaterials/markAsDulyMade",
       "UpdateAccreditationMaterialTaskStatus": "api/v{0}/accreditationMaterialTaskStatus"
