+++ src/EPR.RegulatorService.Facade.Core/Clients/ReprocessorExporter/Accreditations/AccreditationServiceClient.cs
@@ -8,13 +8,13 @@ using Microsoft.Extensions.Options;
 namespace EPR.RegulatorService.Facade.Core.Clients.ReprocessorExporter.Accreditations;
 
 public class AccreditationServiceClient(
-HttpClient httpClient,
-IOptions<PrnBackendServiceApiConfig> options,
-ILogger<AccreditationServiceClient> logger)
-: BaseHttpClient(httpClient), IAccreditationServiceClient
+    HttpClient httpClient,
+    IOptions<PrnBackendServiceApiConfig> options,
+    ILogger<AccreditationServiceClient> logger)
+    : BaseHttpClient(httpClient), IAccreditationServiceClient
 {
     private readonly PrnBackendServiceApiConfig _config = options.Value;
-    
+
     public async Task<AccreditationFeeContextDto> GetAccreditationFeeRequestByRegistrationMaterialId(Guid id)
     {
         logger.LogInformation(LogMessages.AttemptingAccreditationFeeDetails);
@@ -28,12 +28,11 @@ ILogger<AccreditationServiceClient> logger)
         var url = string.Format(_config.Endpoints.AccreditationMarkAsDulyMadeByRegistrationMaterialId, _config.ApiVersion);
         return await PostAsync<AccreditationMarkAsDulyMadeWithUserIdDto, bool>(url, request);
     }
-    
+
     public async Task<bool> UpdateAccreditationMaterialTaskStatus(UpdateAccreditationMaterialTaskStatusWithUserIdDto request)
     {
         logger.LogInformation(LogMessages.UpdateAccreditationMaterialTaskStatus, request.TaskStatus.ToString());
         var url = string.Format(_config.Endpoints.UpdateAccreditationMaterialTaskStatus, _config.ApiVersion);
         return await PostAsync<UpdateAccreditationMaterialTaskStatusDto, bool>(url, request);
     }
-
 }
