+++ src/EPR.RegulatorService.Facade.Core/Clients/ReprocessorExporter/IPaymentServiceClient.cs
@@ -7,5 +7,4 @@ public interface IPaymentServiceClient
     Task<decimal> GetRegistrationPaymentFee(string materialName, string regulator, DateTime submittedDate, string requestorType, string reference);
     Task<bool> SaveOfflinePayment(SaveOfflinePaymentRequestDto request);
     Task<decimal> GetAccreditationPaymentFee(string materialName, string regulator, DateTime submittedDate, string requestorType, string reference);
-
 }
