+++ src/EPR.RegulatorService.Facade.Core/Clients/ReprocessorExporter/PaymentServiceClient.cs
@@ -23,7 +23,6 @@ ILogger<PaymentServiceClient> logger)
         //return await GetAsync<decimal>(url);
     }
 
-    
     public async Task<decimal> GetAccreditationPaymentFee(string materialName, string regulator, DateTime submittedDate, string requestorType, string reference)
     {
         logger.LogInformation(LogMessages.AttemptingAccreditationPaymentFee);
@@ -32,7 +31,6 @@ ILogger<PaymentServiceClient> logger)
         //var url = string.Format($"{_config.Endpoints.GetAccreditationPaymentFee}", _config.ApiVersion, materialName, regulator, submittedDate, requestorType, reference);
         //return await GetAsync<decimal>(url);
     }
-
 
     public async Task<bool> SaveOfflinePayment(SaveOfflinePaymentRequestDto request)
     {
