+++ src/EPR.RegulatorService.Facade.Core/Models/ReprocessorExporter/Accreditations/AccreditationMarkAsDulyMadeWithUserIdDto.cs
@@ -1,7 +1,6 @@
-﻿using System;
-using System.ComponentModel.DataAnnotations;
+﻿using EPR.RegulatorService.Facade.Core.Models.ReprocessorExporter.Registrations;
 
-namespace EPR.RegulatorService.Facade.Core.Models.ReprocessorExporter.Registrations;
+namespace EPR.RegulatorService.Facade.Core.Models.ReprocessorExporter.Accreditations;
 
 public class AccreditationMarkAsDulyMadeWithUserIdDto : AccreditationMarkAsDulyMadeRequestDto
 {
